requirements.txt
fastapi
uvicorn
pika
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-pika
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-propagator-b3
requests

